+++ eco_perf/CMakeLists.txt
@@ -3,8 +3,7 @@ add_library(eco_perf
 )
 set_target_properties(eco_perf PROPERTIES LINKER_LANGUAGE C)
 
-target_link_libraries(eco_perf metrics io graphics)
+target_link_libraries(eco_perf metrics terminal_interface)
 
 add_subdirectory(metrics)
-add_subdirectory(io)
-add_subdirectory(graphics)
+add_subdirectory(terminal_interface)
