--- eco_perf/io/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_library(io
-    format_definitions.h format_definitions.c
-    io.h io.c
-    terminal_cursor.h terminal_cursor.c
-    terminal_interface.h terminal_interface.c
-)
