+++ eco_perf/terminal_interface/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_library(terminal_interface INTERFACE)
+
+target_link_libraries(terminal_interface INTERFACE
+    io graphics terminal)
+
+add_subdirectory(io)
+add_subdirectory(graphics)
+add_subdirectory(terminal)
