+++ eco_perf/terminal_interface/graphics/drawable_percent_bar.h
