+++ eco_perf/terminal_interface/graphics/percent_bar.c
