+++ eco_perf/terminal_interface/graphics/term_layouts.c
