+++ eco_perf/terminal_interface/graphics/terminal_panel.h
