+++ eco_perf/terminal_interface/io/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(io
+    format_definitions.h format_definitions.c
+    io.h io.c
+)
