+++ eco_perf/terminal_interface/terminal/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(terminal
+    cursor.h cursor.c
+    window.h window.c
+)
