+++ eco_perf/terminal_interface/terminal/cursor.h
@@ -1,4 +1,3 @@
-
 #ifndef ECO_PERF_TERMINAL_CURSOR_H_INCLUDED
 #define ECO_PERF_TERMINAL_CURSOR_H_INCLUDED
