+++ test_terminal.c
@@ -1,4 +1,4 @@
-#include "eco_perf/io/io.h"
+#include "eco_perf/terminal_interface/io/io.h"
 
 int main()
 {
